add head and tail matching for ListLike data types
