+++ src/index.ts
@@ -1,9 +1,11 @@
 import { match } from './pattern';
 import { _, end, A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Q, R, S, T, U, V, W, X, Y, Z } from './symbols';
 import { when } from './when';
+import { tail, head } from './keywords';
 
 export {
   match,
   _, A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Q, R, S, T, U, V, W, X, Y, Z,
   when, end,
+  head, tail,
 };
