+++ src/pattern/types.ts
@@ -1,0 +1,1 @@
+export type Pattern = any;
